backend/pyproject.toml
[project]
name = "server-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.127.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0",
    "uvicorn[standard]>=0.40.0",
]
